requirements.txt
google-genai>=0.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
pyyaml>=6.0.1
rich>=13.7.0
pytest>=8.1.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
